pyproject.toml
[project]
name = "langgraph-network-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.56,<0.4.0",
    "langchain-core>=0.2.38,<0.4.0",
    "langchain-google-genai>=1.0.8",
    "langsmith>=0.1.63",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.2",
    "uvicorn>=0.26.0",
    "langgraph-cli[inmem]>=0.1.0",
    "langchain-openai>=0.3.14",
    "mcp[cli]>=1.6.0",
    "langchain-mcp-adapters>=0.0.9",
]
